rename files to include dns
