+++ app/queues/dns/delete-dns-record-flow.server.ts
@@ -4,43 +4,30 @@ import { buildDomain } from '~/utils';
 import { dnsUpdateQueueName } from './workers/dns-update-worker.server';
 import { pollDnsStatusQueueName } from './workers/poll-dns-status-worker.server';
 import { syncDbStatusQueueName } from './workers/sync-db-status-worker.server';
-import { WorkType } from './add-record-flow.server';
-import { updateDnsRecordById } from '~/models/dns-record.server';
+import { WorkType } from './add-dns-record-flow.server';
 
-import type { DnsRecord } from '@prisma/client';
 import type { FlowJob } from 'bullmq';
+import type { DnsRecord } from '@prisma/client';
 import type { DnsUpdaterData } from './workers/dns-update-worker.server';
 import type { DbRecordSynchronizerData } from './workers/sync-db-status-worker.server';
-import type { Subdomain } from './add-record-flow.server';
+import type { Subdomain } from './add-dns-record-flow.server';
 
-export type UpdateDnsRequestData = Pick<DnsRecord, 'id' | 'username' | 'type' | 'value'> &
-  Partial<Pick<DnsRecord, 'description' | 'course' | 'ports'>> &
+export type DeleteDnsRequestData = Pick<DnsRecord, 'id' | 'username' | 'type' | 'value'> &
   Subdomain;
 
 const flowProducer = new FlowProducer({ connection: redis });
 
-export const updateDnsRequest = async (data: UpdateDnsRequestData) => {
-  const { username, type, subdomain, value, id } = data;
+export const deleteDnsRequest = async (data: DeleteDnsRequestData) => {
+  const { id, username, type, subdomain, value } = data;
 
   const fqdn = buildDomain(username, subdomain);
 
-  // Before running workflow, update the dns record in DB
-  await updateDnsRecordById({
-    id: data.id,
-    type: data.type,
-    subdomain,
-    value: data.value,
-    description: data.description,
-    course: data.course,
-    ports: data.ports,
-  });
-
-  // Step 1. Request Route53 to update the record
+  // Step 1. Request Route53 to delete the dns record
   const updateDnsRecord: FlowJob = {
-    name: `updateDnsRecord:${subdomain}-${username}`,
+    name: `deleteDnsRecord:${subdomain}-${username}`,
     queueName: dnsUpdateQueueName,
     data: {
-      workType: WorkType.update,
+      workType: WorkType.delete,
       username,
       type,
       fqdn,
@@ -71,21 +58,21 @@ export const updateDnsRequest = async (data: UpdateDnsRequestData) => {
     },
   };
 
-  // Step 3. Update the MySQL record with the active or error status
-  const syncDbStatus: FlowJob = {
-    name: `syncDbStatus:${subdomain}-${username}`,
+  // Step 3. Delete the record in MySQL
+  const updateDbRecord: FlowJob = {
+    name: `deleteDbRecord:${subdomain}-${username}`,
     queueName: syncDbStatusQueueName,
+    data: { workType: WorkType.delete, id } as DbRecordSynchronizerData,
     children: [pollDnsStatus],
-    data: { workType: WorkType.update, id } as DbRecordSynchronizerData,
     opts: {
       failParentOnFailure: true,
       attempts: 5,
       backoff: {
         type: 'exponential',
-        delay: 30_000,
+        delay: 15_000,
       },
     },
   };
 
-  return flowProducer.add(syncDbStatus);
+  return flowProducer.add(updateDbRecord);
 };
