+++ app/queues/dns/update-dns-record-flow.server.ts
@@ -4,30 +4,43 @@ import { buildDomain } from '~/utils';
 import { dnsUpdateQueueName } from './workers/dns-update-worker.server';
 import { pollDnsStatusQueueName } from './workers/poll-dns-status-worker.server';
 import { syncDbStatusQueueName } from './workers/sync-db-status-worker.server';
-import { WorkType } from './add-record-flow.server';
+import { WorkType } from './add-dns-record-flow.server';
+import { updateDnsRecordById } from '~/models/dns-record.server';
 
-import type { FlowJob } from 'bullmq';
 import type { DnsRecord } from '@prisma/client';
+import type { FlowJob } from 'bullmq';
 import type { DnsUpdaterData } from './workers/dns-update-worker.server';
 import type { DbRecordSynchronizerData } from './workers/sync-db-status-worker.server';
-import type { Subdomain } from './add-record-flow.server';
+import type { Subdomain } from './add-dns-record-flow.server';
 
-export type DeleteDnsRequestData = Pick<DnsRecord, 'id' | 'username' | 'type' | 'value'> &
+export type UpdateDnsRequestData = Pick<DnsRecord, 'id' | 'username' | 'type' | 'value'> &
+  Partial<Pick<DnsRecord, 'description' | 'course' | 'ports'>> &
   Subdomain;
 
 const flowProducer = new FlowProducer({ connection: redis });
 
-export const deleteDnsRequest = async (data: DeleteDnsRequestData) => {
-  const { id, username, type, subdomain, value } = data;
+export const updateDnsRequest = async (data: UpdateDnsRequestData) => {
+  const { username, type, subdomain, value, id } = data;
 
   const fqdn = buildDomain(username, subdomain);
 
-  // Step 1. Request Route53 to delete the dns record
+  // Before running workflow, update the dns record in DB
+  await updateDnsRecordById({
+    id: data.id,
+    type: data.type,
+    subdomain,
+    value: data.value,
+    description: data.description,
+    course: data.course,
+    ports: data.ports,
+  });
+
+  // Step 1. Request Route53 to update the record
   const updateDnsRecord: FlowJob = {
-    name: `deleteDnsRecord:${subdomain}-${username}`,
+    name: `updateDnsRecord:${subdomain}-${username}`,
     queueName: dnsUpdateQueueName,
     data: {
-      workType: WorkType.delete,
+      workType: WorkType.update,
       username,
       type,
       fqdn,
@@ -58,21 +71,21 @@ export const deleteDnsRequest = async (data: DeleteDnsRequestData) => {
     },
   };
 
-  // Step 3. Delete the record in MySQL
-  const updateDbRecord: FlowJob = {
-    name: `deleteDbRecord:${subdomain}-${username}`,
+  // Step 3. Update the MySQL record with the active or error status
+  const syncDbStatus: FlowJob = {
+    name: `syncDbStatus:${subdomain}-${username}`,
     queueName: syncDbStatusQueueName,
-    data: { workType: WorkType.delete, id } as DbRecordSynchronizerData,
     children: [pollDnsStatus],
+    data: { workType: WorkType.update, id } as DbRecordSynchronizerData,
     opts: {
       failParentOnFailure: true,
       attempts: 5,
       backoff: {
         type: 'exponential',
-        delay: 15_000,
+        delay: 30_000,
       },
     },
   };
 
-  return flowProducer.add(updateDbRecord);
+  return flowProducer.add(syncDbStatus);
 };
